+++ swad2moodle.jl
@@ -1,4 +1,5 @@
-using MoodleQuestions
+u
+sing MoodleQuestions
 
 function main(ARGS)
     if length(ARGS)!=2